% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_data.R
\name{extract_data}
\alias{extract_data}
\title{Export data from the Africa Bird Map}
\usage{
extract_data(username, user_id, password, species_id)
}
\arguments{
\item{username}{The citizen scientist's bird map email}

\item{user_id}{The citizen scientist's bird map user number (usually 5 digits)}

\item{password}{The citizen scientist's password.}

\item{species_id}{The species_id for which data is extracted. A complete list of species name and ids are available on the Kenya Bird Map website.}
}
\value{
A dataframe with every list submitted to Africa Bird Map (includes all countries).
}
\description{
Pull data from the African Bird Map fit for reporting rate and other analyses. This function is temporary and will be made redundant. It is unsafe and is only for demonstrative purposes.
}
\details{
The function requires a username and password in plain text. This is not how the final function will look. The final function will call the RStudio API for a safe transaction. This is just for demonstrative purposes.
}
\examples{

\dontrun{

# Extract data for the African Black-shouldered kite.

extract_data(username = "dclarance@gmail.com" ,user_id = 40664, password = "abcd", species_id = 103s)

}


}
