% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_pentads.R
\name{filter_pentads}
\alias{filter_pentads}
\title{Select pentads based on features}
\usage{
filter_pentads(selected_area, selection_type)
}
\arguments{
\item{selected_area}{Either a pentad (eg: 0105_3930), country (eg: Kenya), county (eg: Kitui) or province (eg: rift valley). Lists of the same can be applied as well. For instance, for multiple pentads you could use, c('0105_3930', '0110c3620').}

\item{selection_type}{Can take either of the four values: 'Pentad', 'Country', 'County' or 'Province'.}
}
\value{
A dataframe where each row is a pentade with the `fit` values, the standard errors (`se`), the `reporting_rates` and its associated inputs. Ignore the various date columns as those are only meant to be used in graphs.
}
\description{
Select pentads lying within a defined geographic region. Used as a helper function in the reporting rate functions.
}
\details{
`filter_pentads` is used to generate a character list of pentads associated with a given feature. Currently all features are political (counties, provices, countries). Over time we might consider adding habitat and climate features.


The pentad assignments are correct for all pentads that lie strictly within the national boundaries of Kenya. However since the algorithm uses the midpoint of a pentad to discover its assignment, there are cases where the midpoint lies outside Kenya and so the pentads do not have a classification. In the data these show up as `unclassified`.


Ideally we would want to use an assignment based on area. This will likely show up in the next release.
}
\examples{

# get list of pentads in Rift Valley
filter_pentads(selected_area = "Rift Valley", selection_type = "County")

}
