% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_rate.R
\name{reporting_rate}
\alias{reporting_rate}
\title{Calculate the reporting rate of a species}
\usage{
reporting_rate(df, species_id, start_date = "1970-01-01",
  end_date = Sys.Date(), pentad_id)
}
\arguments{
\item{df}{A dataframe obtained from the species API call. Use `extract_data` if required.}

\item{species_id}{An integer id. The KBM website has a list of all species with their ids.}

\item{start_date}{A character or date object denoting the starting point from which reporting rate is to be measured.}

\item{end_date}{A character or date object denoting the end point from which reporting rate is to be measured.}

\item{pentad_id}{A character vector denoting the pentad id.}
}
\value{
A numeric object between 0 and 1 (inclusive) denoting the reporting rate of a species. The reporting rate is equal to the number of times a species has been recorded divided by the number of full protocol cards in the specified time and region.
}
\description{
Obtain a reporting rate of a species over a specified time period and spatial configuration.
}
\details{
This functions assumes that all years are 365 days. Necessary corrections for leap years are made.
}
\examples{

\dontrun{

# the reporting rate of African Black-shouldered Kite in Nairobi National Park
reporting_rate(df, species_id = 130, start_date = '1970-01-01', end_date = Sys.Date(), pentad_id = '0120-3650')

}


}
