% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/underhill_curves.R
\name{underhill_curves}
\alias{underhill_curves}
\title{Plot the underhill smoother and associated confidence intervals for a species.}
\usage{
underhill_curves(underhill_smoother, species_id, species_name)
}
\arguments{
\item{underhill_smoother}{The dataframe returned by the `underhill_smoother function`.}

\item{species_id}{The species_id for which data is extracted. A complete list of species name and ids are available on the Kenya Bird Map website.}

\item{species_name}{The name of the species you are plotting. This is to use in the title of the plots.}
}
\value{
A ggplot plot
}
\description{
Fit a line through the reporting rates and its associated 1 and 2 sigma confidence intervals. 1 and 2 sigma roughly correspond to 67 percent and 95 percent confidence intervals. The line is over Pentades defined in the `underhill_smoother()` function.
}
\details{
The dark, thinner interval is the 1 sigma interval and corresponds to 67% confidence, whereas the 2 sigma interval corresponds to a 95% confidence.
}
\examples{

\dontrun{


underhill_curves(underhill_smoother, species_id = 103, species_name = "African Black-shouldered Kite")

}


}
