% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/underhill_multiple_curves.R
\name{underhill_multiple_curves}
\alias{underhill_multiple_curves}
\title{Plot the underhill smoother and associated confidence intervals for multiple species.}
\usage{
underhill_multiple_curves(multiple_underhill_smoothers)
}
\arguments{
\item{multiple_underhill_smoothers}{A list of the dataframes returned by the `underhill_smoother function`.}
}
\value{
A ggplot plot
}
\description{
Fit a line through the reporting rates and its associated 2 sigma confidence intervals for multiple species.
}
\details{
To utilize this function first you need to create a list of dataframe. Do this by calling the `list()` function and passing the dataframes as arguments
}
\examples{

\dontrun{

list_of_interesting_species = list(underhill_smoother1 , underhill_smoother2, underhill_smoother3)
underhill_curves(list_of_interesting_species)

}


}
