% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/underhill_single_curves.R
\name{underhill_single_curves}
\alias{underhill_single_curves}
\title{Plot the underhill smoother and associated confidence intervals for a species.}
\usage{
underhill_single_curves(underhill_smoother)
}
\arguments{
\item{underhill_smoother}{The dataframe returned by the `underhill_smoother function`.}
}
\value{
A ggplot plot
}
\description{
Fit a line through the reporting rates and its associated 1 and 2 sigma confidence intervals. 1 and 2 sigma roughly correspond to 67 percent and 95 percent confidence intervals. The line is over Pentades defined in the `underhill_smoother()` function.
}
\details{
The dark, thinner interval is the 1 sigma interval and corresponds to 67% confidence, whereas the 2 sigma interval corresponds to a 95% confidence.

This function only takes in one dataframe and is intended for deep study of one species. To make multi species comparisons use `underhill_multiple_curves`
}
\examples{

\dontrun{


underhill_curves(underhill_smoother)

}


}
