% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/underhill_smoother.R
\name{underhill_smoother}
\alias{underhill_smoother}
\title{Fit the underhill smoother to a reporting rate distribution}
\usage{
underhill_smoother(raw_data, species_id, start_month = 7, pentade_window,
  first_pentade = 1, last_pentade = 73)
}
\arguments{
\item{raw_data}{Data for a species extracted uing `extract_data()`. The StartDate column should be of type `date`.}

\item{species_id}{The species_id for which data is extracted. A complete list of species name and ids are available on the Kenya Bird Map website.}

\item{start_month}{The month to start the pentades from. Default is 7.}

\item{pentade_window}{The number of pentades on either side of the target day to give the weights to.}

\item{first_pentade}{The starting value of the pentades. Default is 1.}

\item{last_pentade}{The last value of the pentades. Default is 73.}
}
\value{
A dataframe where each row is a pentade with the `fit` values, the standard errors (`se`), the `reporting_rates` and its associated inputs. Ignore the various date columns as those are only meant to be used in graphs.
}
\description{
Fit the underhill smoother (put reference here) to reporting rate distribution over pentades.The underhill smoother is a locally weighted binomial general linear model using a logit link, where the weights are generated using a exponential distribution.
}
\details{
Pentades are blocks of 5 consecutive days. The implementation measures pentades from a defined month `start_month`. This implementation needs review and should not be considered final.
}
\examples{

\dontrun{


underhill_smoother(raw_data, species_id, start_month = 7, pentade_window, first_pentade = 1, last_pentade = 73)

}


}
